Set key field on Insert when possible
The ability to have auto-incrementing keys is really useful, but becomes
a lot less useful if you need to use that generated key as a reference
somewhere else because you can only get it by performing a (likely
expensive) query to retrieve the value that was just inserted.

To support this scenario, which includes supporting roll-your-own
foreign keys, have `Insert()` write the actual key used in to the data
when certain criteria are met:
 - the key type and the field type must be the same
 - the value in the data must be the zero-value of that type
 - the data must be passed in by reference

For scenarios where this behaviour is not wanted it is generally simply
a matter of passing the data by value instead of by pointer.