improve query handling on empty buckets
If the bucket for the storer doesn't exist or is empty then we can
bypass all of the usual work and just immediately return.

In addition to being faster, this also allows us to return an empty
result set instead of an error when attempting to use an index before it
has been created.

Fixes #30